+++ code/engine.vc2008/xrEngine/bone.h
@@ -485,9 +485,9 @@ private:
 public:
 	virtual u32			mem_usage		()
 	{
-		u32 sz			= sizeof(*this)+sizeof(vecBones::value_type)*children.size();
+		u32 sz			= sizeof(*this)+sizeof(vecBones::value_type)*u32(children.size());
 		for (auto c_it=child_faces.begin(); c_it!=child_faces.end(); c_it++)
-			sz			+= c_it->size()*sizeof(FacesVec::value_type)+sizeof(*c_it);
+			sz			+= u32(c_it->size()*sizeof(FacesVec::value_type)+sizeof(*c_it));
 		return			sz;
 	}
 };
